Add files via upload
Converted PodGrab.py to work with Python3.72 using requests instead of urllib, and added a feature for custom headers (for User Agent) so downloads are a bit more friendly from the RSS feeds that otherwise block your downloads when they think it's a scraper or bot. This mod now picks up all the podcast the original didn't in my tests, but it's still a bit buggy in some of the Python 2 to 3 conversion spots.